Add new components and environment configurations
This commit includes the creation of new components for modal, header, footer, and welcome. It also contains the integration of environment configurations for production and development. Furthermore, existing unnecessary files, including app.component.spec.ts and vscode configurations, have been removed. This update is expected to organize and optimize the structure of the project.