Add 404 not found page and its translations
A new error 404 page and its corresponding translation files have been added. This is implemented to handle and provide relevant messages to the user when they navigate to a non-existing page. The added files are for the English and Russian versions of the error messaging. This includes the implementation of a 'Go Back' button for redirecting users back to the homepage.